+++ pyimgur/__init__.py
@@ -48,10 +48,13 @@ from pyimgur import request  # NOQA
 
 __version__ = '0.5.3'
 
-AUTHORIZE_URL = ("https://api.imgur.com/oauth2/authorize?"
+MASHAPE_BASE = "https://imgur-apiv3.p.mashape.com"
+IMGUR_BASE = "https://api.imgur.com"
+
+AUTHORIZE_URL = ("{}/oauth2/authorize?"
                  "client_id={}&response_type={}&state={}")
-EXCHANGE_URL = "https://api.imgur.com/oauth2/token"
-REFRESH_URL = "https://api.imgur.com/oauth2/token"
+EXCHANGE_URL = "{}/oauth2/token"
+REFRESH_URL = "{}/oauth2/token"
 
 
 def _change_object(from_object, to_object):
@@ -254,7 +257,7 @@ class Album(Basic_object):
     :ivar views: Total number of views the album has received.
     """
     def __init__(self, json_dict, imgur, has_fetched=True):
-        self._INFO_URL = ("https://api.imgur.com/3/album/"
+        self._INFO_URL = (imgur._base_url + "/3/album/"
                           "{0}".format(json_dict['id']))
         self.deletehash = None
         super(Album, self).__init__(json_dict, imgur, has_fetched)
@@ -268,14 +271,14 @@ class Album(Basic_object):
             cannot add (non-existing or not owned by you) will not cause
             exceptions, but fail silently.
         """
-        url = "https://api.imgur.com/3/album/{0}/add".format(self.id)
+        url = self._imgur._base_url + "/3/album/{0}/add".format(self.id)
         params = {'ids': images}
         return self._imgur._send_request(url, needs_auth=True, params=params,
                                          method="POST")
 
     def delete(self):
         """Delete this album."""
-        url = ("https://api.imgur.com/3/album/"
+        url = (self._imgur._base_url + "/3/album/"
                "{0}".format(self._delete_or_id_hash))
         return self._imgur._send_request(url, method="DELETE")
 
@@ -285,7 +288,7 @@ class Album(Basic_object):
 
         Favoriting an already favorited album will unfavor it.
         """
-        url = "https://api.imgur.com/3/album/{0}/favorite".format(self.id)
+        url = self._imgur._base_url + "/3/album/{0}/favorite".format(self.id)
         return self._imgur._send_request(url, needs_auth=True, method="POST")
 
     def remove_images(self, images):
@@ -297,7 +300,7 @@ class Album(Basic_object):
             you cannot remove (non-existing, not owned by you or not part of
             album) will not cause exceptions, but fail silently.
         """
-        url = ("https://api.imgur.com/3/album/{0}/"
+        url = (self._imgur._base_url + "/3/album/{0}/"
                "remove_images".format(self._delete_or_id_hash))
         # NOTE: Returns True and everything seem to be as it should in testing.
         # Seems most likely to be upstream bug.
@@ -313,7 +316,7 @@ class Album(Basic_object):
             images that you cannot set (non-existing or not owned by you) will
             not cause exceptions, but fail silently.
         """
-        url = ("https://api.imgur.com/3/album/"
+        url = (self._imgur._base_url + "/3/album/"
                "{0}/".format(self._delete_or_id_hash))
         params = {'ids': images}
         return self._imgur._send_request(url, needs_auth=True, params=params,
@@ -331,7 +334,7 @@ class Album(Basic_object):
             this method will return an error. Set this to True to by-pass the
             terms.
         """
-        url = "https://api.imgur.com/3/gallery/{0}".format(self.id)
+        url = self._imgur._base_url + "/3/gallery/{0}".format(self.id)
         payload = {'title': title, 'terms': '1' if bypass_terms else '0'}
         self._imgur._send_request(url, needs_auth=True, params=payload,
                                   method='POST')
@@ -358,7 +361,7 @@ class Album(Basic_object):
         :param layout: The way the album is displayed, can be blog, grid,
             horizontal or vertical.
         """
-        url = ("https://api.imgur.com/3/album/"
+        url = (self._imgur._base_url + "/3/album/"
                "{0}".format(self._delete_or_id_hash))
         is_updated = self._imgur._send_request(url, params=locals(),
                                                method='POST')
@@ -409,13 +412,13 @@ class Comment(Basic_object):
     """
     def __init__(self, json_dict, imgur, has_fetched=True):
         self.deletehash = None
-        self._INFO_URL = ("https://api.imgur.com/3/comment/"
+        self._INFO_URL = (imgur._base_url + "/3/comment/"
                           "{0}".format(json_dict['id']))
         super(Comment, self).__init__(json_dict, imgur, has_fetched)
 
     def delete(self):
         """Delete the comment."""
-        url = ("https://api.imgur.com/3/image/"
+        url = (self._imgur._base_url + "/3/image/"
                "{0}".format(self._delete_or_id_hash))
         return self._imgur._send_request(url, method='DELETE')
         # NOTE: Gives a 403 permission denied error on comment 77087313 which
@@ -423,19 +426,19 @@ class Comment(Basic_object):
 
     def downvote(self):
         """Downvote this comment."""
-        url = "https://api.imgur.com/3/comment/{0}/vote/down".format(self.id)
+        url = self._imgur._base_url + "/3/comment/{0}/vote/down".format(self.id)
         return self._imgur._send_request(url, needs_auth=True, method='POST')
 
     def get_replies(self):
         """Get the replies to this comment."""
-        url = "https://api.imgur.com/3/comment/{0}/replies".format(self.id)
+        url = self._imgur._base_url + "/3/comment/{0}/replies".format(self.id)
         json = self._imgur._send_request(url)
         child_comments = json['children']
         return [Comment(com, self._imgur) for com in child_comments]
 
     def reply(self, text):
         """Make a comment reply."""
-        url = "https://api.imgur.com/3/comment/{0}".format(self.id)
+        url = self._imgur._base_url + "/3/comment/{0}".format(self.id)
         payload = {'image_id': self.image.id, 'comment': text}
         resp = self._imgur._send_request(url, params=payload, needs_auth=True,
                                          method='POST')
@@ -452,7 +455,7 @@ class Comment(Basic_object):
 
     def upvote(self):
         """Upvote this comment."""
-        url = "https://api.imgur.com/3/comment/{0}/vote/up".format(self.id)
+        url = self._imgur._base_url + "/3/comment/{0}/vote/up".format(self.id)
         return self._imgur._send_request(url, needs_auth=True, method='POST')
 
 
@@ -464,7 +467,7 @@ class Gallery_item(object):
 
         :param text: The comment text.
         """
-        url = "https://api.imgur.com/3/comment"
+        url = self._imgur._base_url + "/3/comment"
         payload = {'image_id': self.id, 'comment': text}
         resp = self._imgur._send_request(url, params=payload, needs_auth=True,
                                          method='POST')
@@ -478,18 +481,18 @@ class Gallery_item(object):
         something the authenticated user has already downvoted will set the
         vote to neutral.
         """
-        url = "https://api.imgur.com/3/gallery/{0}/vote/down".format(self.id)
+        url = self._imgur._base_url + "/3/gallery/{0}/vote/down".format(self.id)
         return self._imgur._send_request(url, needs_auth=True, method='POST')
 
     def get_comments(self):
         """Get a list of the top-level comments."""
-        url = "https://api.imgur.com/3/gallery/{0}/comments".format(self.id)
+        url = self._imgur._base_url + "/3/gallery/{0}/comments".format(self.id)
         resp = self._imgur._send_request(url)
         return [Comment(com, self._imgur) for com in resp]
 
     def remove_from_gallery(self):
         """Remove this image from the gallery."""
-        url = "https://api.imgur.com/3/gallery/{0}".format(self.id)
+        url = self._imgur._base_url + "/3/gallery/{0}".format(self.id)
         self._imgur._send_request(url, needs_auth=True, method='DELETE')
         if isinstance(self, Image):
             item = self._imgur.get_image(self.id)
@@ -506,7 +509,7 @@ class Gallery_item(object):
         something the authenticated user has already upvoted will set the vote
         to neutral.
         """
-        url = "https://api.imgur.com/3/gallery/{0}/vote/up".format(self.id)
+        url = self._imgur._base_url + "/3/gallery/{0}/vote/up".format(self.id)
         return self._imgur._send_request(url, needs_auth=True, method='POST')
 
 
@@ -537,14 +540,14 @@ class Image(Basic_object):
     :ivar width: The width of the image in bytes.
     """
     def __init__(self, json_dict, imgur, has_fetched=True):
-        self._INFO_URL = ("https://api.imgur.com/3/image/"
+        self._INFO_URL = (imgur._base_url + "/3/image/"
                           "{0}".format(json_dict['id']))
         self.deletehash = None
         super(Image, self).__init__(json_dict, imgur, has_fetched)
 
     def delete(self):
         """Delete the image."""
-        url = ("https://api.imgur.com/3/image/"
+        url = (self._imgur._base_url + "/3/image/"
                "{0}".format(self._delete_or_id_hash))
         return self._imgur._send_request(url, method='DELETE')
 
@@ -602,7 +605,7 @@ class Image(Basic_object):
 
         Favoriting an already favorited image will unfavorite it.
         """
-        url = "https://api.imgur.com/3/image/{0}/favorite".format(self.id)
+        url = self._imgur._base_url + "/3/image/{0}/favorite".format(self.id)
         return self._imgur._send_request(url, needs_auth=True, method='POST')
 
     def submit_to_gallery(self, title, bypass_terms=False):
@@ -617,7 +620,7 @@ class Image(Basic_object):
             this method will return an error. Set this to True to by-pass the
             terms.
         """
-        url = "https://api.imgur.com/3/gallery/{0}".format(self.id)
+        url = self._imgur._base_url + "/3/gallery/{0}".format(self.id)
         payload = {'title': title, 'terms': '1' if bypass_terms else '0'}
         self._imgur._send_request(url, needs_auth=True, params=payload,
                                   method='POST')
@@ -627,7 +630,7 @@ class Image(Basic_object):
 
     def update(self, title=None, description=None):
         """Update the image with a new title and/or description."""
-        url = ("https://api.imgur.com/3/image/"
+        url = (self._imgur._base_url + "/3/image/"
                "{0}".format(self._delete_or_id_hash))
         is_updated = self._imgur._send_request(url, params=locals(),
                                                method='POST')
@@ -646,7 +649,7 @@ class Imgur:
     but instead use the methods in this class to get them.
     """
     def __init__(self, client_id, client_secret=None, access_token=None,
-                 refresh_token=None, verify=True):
+                 refresh_token=None, verify=True, mashape_key=None):
         """
         Initialize the Imgur object.
 
@@ -680,6 +683,11 @@ class Imgur:
         self.ratelimit_userreset = None
         self.refresh_token = refresh_token
         self.verify = verify
+        self.mashape_key = mashape_key
+        if self.mashape_key:
+            self._base_url = MASHAPE_BASE
+        else:
+            self._base_url = IMGUR_BASE
 
     def _send_request(self, url, needs_auth=False, **kwargs):
         """
@@ -707,6 +715,8 @@ class Imgur:
             auth = {'Authorization': 'Client-ID {0}'.format(self.client_id)}
         else:
             auth = {'Authorization': 'Bearer {0}'.format(self.access_token)}
+        if self.mashape_key:
+            auth.update({'X-Mashape-Key': self.mashape_key})
         content = []
         is_paginated = False
         if 'limit' in kwargs:
@@ -755,7 +765,7 @@ class Imgur:
             correct resource in your site, nonces, and
             cross-site-request-forgery mitigations.
         """
-        return AUTHORIZE_URL.format(self.client_id, response, state)
+        return AUTHORIZE_URL.format(self._base_url, self.client_id, response, state)
 
     def change_authentication(self, client_id=None, client_secret=None,
                               access_token=None, refresh_token=None):
@@ -782,7 +792,7 @@ class Imgur:
 
         :returns: The newly created album.
         """
-        url = "https://api.imgur.com/3/album/"
+        url = self._base_url + "/3/album/"
         payload = {'ids': images, 'title': title,
                    'description': description, 'cover': cover}
         resp = self._send_request(url, params=payload, method='POST')
@@ -806,7 +816,8 @@ class Imgur:
                   'client_secret': self.client_secret,
                   'grant_type': 'authorization_code',
                   'code': code}
-        result = self._send_request(EXCHANGE_URL, params=params, method='POST',
+        result = self._send_request(EXCHANGE_URL.format(self._base_url),
+                                    params=params, method='POST',
                                     data_field=None)
         self.access_token = result['access_token']
         self.refresh_token = result['refresh_token']
@@ -818,7 +829,8 @@ class Imgur:
                   'client_secret': self.client_secret,
                   'grant_type': 'pin',
                   'pin': pin}
-        result = self._send_request(EXCHANGE_URL, params=params, method='POST',
+        result = self._send_request(EXCHANGE_URL.format(self._base_url),
+                                    params=params, method='POST',
                                     data_field=None)
         self.access_token = result['access_token']
         self.refresh_token = result['refresh_token']
@@ -826,7 +838,7 @@ class Imgur:
 
     def get_album(self, id):
         """Return information about this album."""
-        url = "https://api.imgur.com/3/album/{0}".format(id)
+        url = self._base_url + "/3/album/{0}".format(id)
         json = self._send_request(url)
         return Album(json, self)
 
@@ -928,7 +940,7 @@ class Imgur:
 
     def get_comment(self, id):
         """Return information about this comment."""
-        url = "https://api.imgur.com/3/comment/{0}".format(id)
+        url = self._base_url + "/3/comment/{0}".format(id)
         json = self._send_request(url)
         return Comment(json, self)
 
@@ -945,7 +957,7 @@ class Imgur:
             'user' section. Defaults to true.
         :param limit: The number of items to return.
         """
-        url = ("https://api.imgur.com/3/gallery/{}/{}/{}/{}?showViral="
+        url = (self._base_url + "/3/gallery/{}/{}/{}/{}?showViral="
                "{}".format(section, sort, window, '{}', show_viral))
         resp = self._send_request(url, limit=limit)
         return [_get_album_or_image(thing, self) for thing in resp]
@@ -958,7 +970,7 @@ class Imgur:
         This makes it possible to remove an album from the gallery and setting
         it's privacy setting as secret, without compromising it's secrecy.
         """
-        url = "https://api.imgur.com/3/gallery/album/{0}".format(id)
+        url = self._base_url + "/3/gallery/album/{0}".format(id)
         resp = self._send_request(url)
         return Gallery_album(resp, self)
 
@@ -970,13 +982,13 @@ class Imgur:
         This makes it possible to remove an image from the gallery and setting
         it's privacy setting as secret, without compromising it's secrecy.
         """
-        url = "https://api.imgur.com/3/gallery/image/{0}".format(id)
+        url = self._base_url + "/3/gallery/image/{0}".format(id)
         resp = self._send_request(url)
         return Gallery_image(resp, self)
 
     def get_image(self, id):
         """Return a Image object representing the image with the given id."""
-        url = "https://api.imgur.com/3/image/{0}".format(id)
+        url = self._base_url + "/3/image/{0}".format(id)
         resp = self._send_request(url)
         return Image(resp, self)
 
@@ -986,7 +998,7 @@ class Imgur:
 
         :param id: The id of the message object to return.
         """
-        url = "https://api.imgur.com/3/message/{0}".format(id)
+        url = self._base_url + "/3/message/{0}".format(id)
         resp = self._send_request(url)
         return Message(resp, self)
 
@@ -996,7 +1008,7 @@ class Imgur:
 
         :param id: The id of the notification object to return.
         """
-        url = "https://api.imgur.com/3/notification/{0}".format(id)
+        url = self._base_url + "/3/notification/{0}".format(id)
         resp = self._send_request(url)
         return Notification(resp, self)
 
@@ -1011,7 +1023,7 @@ class Imgur:
             "top", day | week | month | year | all, defaults to week.
         :param limit: The number of items to return.
         """
-        url = ("https://api.imgur.com/3/gallery/g/memes/{0}/{1}/{2}".format(
+        url = (self._base_url + "/3/gallery/g/memes/{0}/{1}/{2}".format(
                sort, window, '{}'))
         resp = self._send_request(url, limit=limit)
         return [_get_album_or_image(thing, self) for thing in resp]
@@ -1028,7 +1040,7 @@ class Imgur:
 
         :param id: The id of the image we want.
         """
-        url = "https://api.imgur.com/3/gallery/g/memes/" % id
+        url = self._base_url + "/3/gallery/g/memes/" % id
         resp = self._send_request(url)
         return Gallery_image(resp, self)
     '''
@@ -1047,7 +1059,7 @@ class Imgur:
             "top", day | week | month | year | all, defaults to day.
         :param limit: The number of items to return.
         """
-        url = ("https://api.imgur.com/3/gallery/r/{0}/{1}/{2}/{3}".format(
+        url = (self._base_url + "/3/gallery/r/{0}/{1}/{2}/{3}".format(
                subreddit, sort, window, '{}'))
         resp = self._send_request(url, limit=limit)
         return [_get_album_or_image(thing, self) for thing in resp]
@@ -1059,7 +1071,7 @@ class Imgur:
         :param subreddit: The subreddit the image has been submitted to.
         :param id: The id of the image we want.
         """
-        url = "https://api.imgur.com/3/gallery/r/{0}/{1}".format(subreddit, id)
+        url = self._base_url + "/3/gallery/r/{0}/{1}".format(subreddit, id)
         resp = self._send_request(url)
         return Gallery_image(resp, self)
 
@@ -1069,7 +1081,7 @@ class Imgur:
 
         :param username: The name of the user we want more information about.
         """
-        url = "https://api.imgur.com/3/account/{0}".format(username)
+        url = self._base_url + "/3/account/{0}".format(username)
         json = self._send_request(url)
         return User(json, self)
 
@@ -1094,14 +1106,15 @@ class Imgur:
                   'client_secret': self.client_secret,
                   'grant_type': 'refresh_token',
                   'refresh_token': self.refresh_token}
-        result = self._send_request(REFRESH_URL, params=params, method='POST',
+        result = self._send_request(REFRESH_URL.format(self._base_url),
+                                    params=params, method='POST',
                                     data_field=None)
         self.access_token = result['access_token']
         return self.access_token
 
     def search_gallery(self, q):
         """Search the gallery with the given query string."""
-        url = "https://api.imgur.com/3/gallery/search?q={0}".format(q)
+        url = self._base_url + "/3/gallery/search?q={0}".format(q)
         resp = self._send_request(url)
         return [_get_album_or_image(thing, self) for thing in resp]
 
@@ -1134,7 +1147,7 @@ class Imgur:
         payload = {'album_id': album, 'image': image,
                    'title': title, 'description': description}
 
-        resp = self._send_request("https://api.imgur.com/3/image",
+        resp = self._send_request(self._base_url + "/3/image",
                                   params=payload, method='POST')
         # TEMPORARY HACK:
         # On 5-08-2013 I noticed Imgur now returned enough information from
@@ -1152,7 +1165,7 @@ class Imgur:
 class Message(Basic_object):
     """This corresponds to the messages users can send each other."""
     def __init__(self, json_dict, imgur, has_fetched=True):
-        self._INFO_URL = ("https://api.imgur.com/3/message/{0}".format(
+        self._INFO_URL = (imgur._base_url + "/3/message/{0}".format(
                           json_dict['id']))
         super(Message, self).__init__(json_dict, imgur, has_fetched)
 
@@ -1168,12 +1181,12 @@ class Message(Basic_object):
 
     def delete(self):
         """Delete the message."""
-        url = "https://api.imgur.com/3/message/{0}".format(self.id)
+        url = self._imgur._base_url + "/3/message/{0}".format(self.id)
         return self._imgur._send_request(url, method='DELETE')
 
     def get_thread(self):
         """Return the message thread this Message is in."""
-        url = ("https://api.imgur.com/3/message/{0}/thread".format(
+        url = (self._imgur._base_url + "/3/message/{0}/thread".format(
                self.first_message.id))
         resp = self._imgur._send_request(url)
         return [Message(msg, self._imgur) for msg in resp]
@@ -1215,7 +1228,7 @@ class Notification(Basic_object):
     """
     def __init__(self, json_dict, imgur, has_fetched=True):
         # Is never gotten lazily, so _has_fetched is always True
-        self._INFO_URL = ("https://api.imgur.com/3/notification/{0}".format(
+        self._INFO_URL = (imgur._base_url + "/3/notification/{0}".format(
                           json_dict['id']))
         super(Notification, self).__init__(json_dict, imgur, has_fetched)
 
@@ -1225,7 +1238,7 @@ class Notification(Basic_object):
 
         Notifications cannot be marked as unviewed.
         """
-        url = "https://api.imgur.com/3/notification/{0}".format(self.id)
+        url = self._imgur._base_url + "/3/notification/{0}".format(self.id)
         return self._imgur._send_request(url, method='POST')
 
 
@@ -1241,7 +1254,7 @@ class User(Basic_object):
     :ivar reputation: Total likes - dislikes of the user's created content.
     """
     def __init__(self, json_dict, imgur, has_fetched=True):
-        self._INFO_URL = ("https://api.imgur.com/3/account/{0}".format(
+        self._INFO_URL = (imgur._base_url + "/3/account/{0}".format(
                           json_dict['url']))
         super(User, self).__init__(json_dict, imgur, has_fetched)
 
@@ -1268,14 +1281,14 @@ class User(Basic_object):
         # NOTE: album_privacy should maybe be renamed to default_privacy
         # NOTE: public_images is a boolean, despite the documentation saying it
         # is a string.
-        url = "https://api.imgur.com/3/account/{0}/settings".format(self.name)
+        url = self._imgur._base_url + "/3/account/{0}/settings".format(self.name)
         resp = self._imgur._send_request(url, needs_auth=True, params=locals(),
                                          method='POST')
         return resp
 
     def delete(self):
         """Delete this user. Require being authenticated as the user."""
-        url = "https://api.imgur.com/3/account/{0}".format(self.name)
+        url = self._imgur._base_url + "/3/account/{0}".format(self.name)
         return self._imgur._send_request(url, needs_auth=True, method='DELETE')
 
     def get_albums(self, limit=None):
@@ -1285,33 +1298,33 @@ class User(Basic_object):
         Secret and hidden albums are only returned if this is the logged-in
         user.
         """
-        url = ("https://api.imgur.com/3/account/{0}/albums/{1}".format(self.name,
+        url = (self._imgur._base_url + "/3/account/{0}/albums/{1}".format(self.name,
                                                                        '{}'))
         resp = self._imgur._send_request(url, limit=limit)
         return [Album(alb, self._imgur, False) for alb in resp]
 
     def get_comments(self):
         """Return the comments made by the user."""
-        url = "https://api.imgur.com/3/account/{0}/comments".format(self.name)
+        url = self._imgur._base_url + "/3/account/{0}/comments".format(self.name)
         resp = self._imgur._send_request(url)
         return [Comment(com, self._imgur) for com in resp]
 
     def get_favorites(self):
         """Return the users favorited images."""
-        url = "https://api.imgur.com/3/account/{0}/favorites".format(self.name)
+        url = self._imgur._base_url + "/3/account/{0}/favorites".format(self.name)
         resp = self._imgur._send_request(url, needs_auth=True)
         return [_get_album_or_image(thing, self) for thing in resp]
 
     def get_gallery_favorites(self):
         """Get a list of the images in the gallery this user has favorited."""
-        url = ("https://api.imgur.com/3/account/{0}/gallery_favorites".format(
+        url = (self._imgur._base_url + "/3/account/{0}/gallery_favorites".format(
                self.name))
         resp = self._imgur._send_request(url)
         return [Image(img, self._imgur) for img in resp]
 
     def get_gallery_profile(self):
         """Return the users gallery profile."""
-        url = ("https://api.imgur.com/3/account/{0}/"
+        url = (self._imgur._base_url + "/3/account/{0}/"
                "gallery_profile".format(self.name))
         return self._imgur._send_request(url)
 
@@ -1323,13 +1336,13 @@ class User(Basic_object):
         sending an email to the user and the owner of the email user verifying
         that he is the same as the Imgur user.
         """
-        url = ("https://api.imgur.com/3/account/{0}/"
+        url = (self._imgur._base_url + "/3/account/{0}/"
                "verifyemail".format(self.name))
         return self._imgur._send_request(url, needs_auth=True)
 
     def get_images(self, limit=None):
         """Return all of the images associated with the user."""
-        url = ("https://api.imgur.com/3/account/{0}/"
+        url = (self._imgur._base_url + "/3/account/{0}/"
                "images/{1}".format(self.name, '{}'))
         resp = self._imgur._send_request(url, limit=limit)
         return [Image(img, self._imgur) for img in resp]
@@ -1341,7 +1354,7 @@ class User(Basic_object):
         :param new: False for all notifications, True for only non-viewed
             notifications.
         """
-        url = ("https://api.imgur.com/3/account/{0}/notifications/"
+        url = (self._imgur._base_url + "/3/account/{0}/notifications/"
                "messages".format(self.name))
         result = self._imgur._send_request(url, params=locals(),
                                            needs_auth=True)
@@ -1350,7 +1363,7 @@ class User(Basic_object):
 
     def get_notifications(self, new=True):
         """Return all the notifications for this user."""
-        url = ("https://api.imgur.com/3/account/{0}/"
+        url = (self._imgur._base_url + "/3/account/{0}/"
                "notifications".format(self.name))
         resp = self._imgur._send_request(url, params=locals(), needs_auth=True)
         msgs = [Message(msg_dict, self._imgur, has_fetched=True) for msg_dict
@@ -1366,7 +1379,7 @@ class User(Basic_object):
         :param new: False for all notifications, True for only non-viewed
             notifications.
         """
-        url = ("https://api.imgur.com/3/account/{0}/"
+        url = (self._imgur._base_url + "/3/account/{0}/"
                "notifications/replies".format(self.name))
         return self._imgur._send_request(url, needs_auth=True)
 
@@ -1376,17 +1389,17 @@ class User(Basic_object):
 
         Only accessible if authenticated as the user.
         """
-        url = "https://api.imgur.com/3/account/{0}/settings".format(self.name)
+        url = self._imgur._base_url + "/3/account/{0}/settings".format(self.name)
         return self._imgur._send_request(url)
 
     def get_statistics(self):
         """Return statistics about this user."""
-        url = "https://api.imgur.com/3/account/{0}/stats".format(self.name)
+        url = self._imgur._base_url + "/3/account/{0}/stats".format(self.name)
         return self._imgur._send_request(url, needs_auth=True)
 
     def get_submissions(self, limit=None):
         """Return a list of the images a user has submitted to the gallery."""
-        url = ("https://api.imgur.com/3/account/{0}/submissions/"
+        url = (self._imgur._base_url + "/3/account/{0}/submissions/"
                "{1}".format(self.name, '{}'))
         resp = self._imgur._send_request(url, limit=limit)
         return [_get_album_or_image(thing, self._imgur) for thing in resp]
@@ -1404,7 +1417,7 @@ class User(Basic_object):
             message thread. If it's a Message object or message_id, then the
             new message will be sent as a reply to the reply_to message.
         """
-        url = "https://api.imgur.com/3/message"
+        url = self._imgur._base_url + "/3/message"
         parent_id = reply_to.id if isinstance(reply_to, Message) else reply_to
         payload = {'recipient': self.name, 'body': body, 'subject': subject,
                    'parent_id': parent_id}
@@ -1418,7 +1431,7 @@ class User(Basic_object):
         Remember that the verification email may end up in the users spam
         folder.
         """
-        url = ("https://api.imgur.com/3/account/{0}"
+        url = (self._imgur._base_url + "/3/account/{0}"
                "/verifyemail".format(self.name))
         self._imgur._send_request(url, needs_auth=True, method='POST')
 
@@ -1429,7 +1442,7 @@ class User(Basic_object):
 class Gallery_album(Album, Gallery_item):
     """Gallery Albums are albums submitted to the gallery."""
     def __init__(self, json_dict, imgur, has_fetched=True):
-        self._INFO_URL = ("https://api.imgur.com/3/gallery/album/"
+        self._INFO_URL = (imgur._base_url + "/3/gallery/album/"
                           "{0}".format(json_dict['id']))
         super(Gallery_album, self).__init__(json_dict, imgur, has_fetched)
 
@@ -1437,6 +1450,6 @@ class Gallery_album(Album, Gallery_item):
 class Gallery_image(Image, Gallery_item):
     """Gallery images are images submitted to the gallery."""
     def __init__(self, json_dict, imgur, has_fetched=True):
-        self._INFO_URL = ("http://api.imgur.com/3/gallery/image/"
+        self._INFO_URL = (imgur._base_url + "/3/gallery/image/"
                           "{0}".format(json_dict['id']))
         super(Gallery_image, self).__init__(json_dict, imgur, has_fetched)