+++ CHANGES.rst
@@ -20,6 +20,8 @@ Unreleased
    than the expected `AttributeError`.
  * **[CHANGE]** Tests that require authentication will now be skipped if
    authentication has not been set up.
+ * **[FEATURE]** Add support for Mashape API Keys for commercial usage of the Imgur API.
+   See: http://api.imgur.com/#commercial
 
 PyImgur 0.5.3
 -------------